Work based on this slack missive:
Jinal Shah [8:56 AM]
@channel: So, had a thought last night.
We need to change the versioning of AMI project repos to a simple build timestamp
- now that's you've enabled triggered AMI builds, there is too large a chance of
a race condition:

A second build could be triggered within the 10 minutes it takes to build the first.
As the first has not completed successfully yet, it has not pushed the version tag
to the git repo, so both builds would have the same git tag.
If both builds finish successfully, you'd end up with two different AMI both marked
with the same tag but the second one would contain a different (more recent) jar.
The second job would also look like a failure as it would not be able to push its
duplicate tag to the repo. This will be as confusing as big, nasty ballz.

The new workflow (only for automated project tagging - doesn't affect devs' workflow):
1) the make command will look for local timestamp tags that don't exist on the remote
and delete them. This accounts for the local dev pattern - trying a build and aborting
it or other failure conditions.

2) Next it will (only if there are no uncommitted changes) git-tag locally with the
current timestamp.

3) On successful build this will be pushed to the remote repo by the make workflow,
not jenkins. i.e. tagging AMI project version (not app_version) is the responsibility
of Make.

If anyone objects, scream. Loudly.

Please note *we will still use semver (preferred because of the information it conveys
about backwards-breaking changes) on all included repos such as packer_includes.*
These follow a different lifecycle and are not subject to such race conditions

Additionally this will rely on us only marking an AMI as stable if we can confirm that
the sha and tag match with the aws ami tag information.